requirements.txt
aiogram==3.22.0
APScheduler==3.11.0
aiosqlite==0.21.0
Pillow==11.3.0
torch==2.8.0
torchvision==0.23.0
